Travis yml: update macOS, lcov and drop brew update.

+++ include/bitcoin/database/databases/block_database.hpp
@@ -46,7 +46,7 @@ public:
         const path& confirmed_index_filename, const path& tx_index_filename,
         size_t table_minimum, size_t candidate_index_minimum,
         size_t confirmed_index_minimum, size_t tx_index_minimum,
-        size_t buckets, size_t expansion, bool neutrino_filters);
+        uint32_t buckets, size_t expansion, bool neutrino_filters);
 
     /// Close the database (all threads must first be stopped).
     ~block_database();
